Do not init() Project on stopProject()
Some CuteElements might get broken during execution of the project.
The user might want to stop the project if such thing will occur and
inspect what's wrong.
And if the user will stop the Project, init() will reset statuses
and health of all CuteElements so the user will not be able to locate
the problem.