Fix one TextField in RunProgramAction controller
workingDirTextField in RunProgramActionController is broken.
It's being erased on each change of pathTextField causing
buttonStateManager to think that the user has changed
workingDirTextField and changes can be applied to the model.

This commit will stop erase of workingDirTextField on each change
of pathTextField.
Also it will do some clean up in RunProgramAction.java and
RunProgramActionController.java.