StreamSis Service Pack 1
A big shameful commit:
1. Fix many architectural problems in GUI related to editing CuteElements.
Editing panels are now working on CuteElement's copy.
2. Rename ApplyAndOkButtonsStateManager to CuteButtonsStatesManager and make it
work as intended.
3. Change the way of serialization of CuteElements to allow easily create
copies of CuteElements in GUI. Each serialized CuteElement now contains full
name of it's class, which is not good, because it will make program less
flexible if we want to maintain backward compatibility in the future.