Make "Perform Test" button work
This patch makes "Perform Test" button work when editing Actions,
Checkers and Counters. It makes the button disabled when editing other
types of CuteElements.

The test is performed in a background thread and is followed by fancy
animations. During the test "Perform Button" transforms to show testing
status and testing result. Testing result for Actions: "Done", for
Checkers: True or False, for Counters: a number.