Improve SisScene and SisSceneCell
Shameful patch that would be better if it was split into smaller
patches. It does:
1. Rename "Default" SisScene to "Primary" SisScene. It sounds cooler.
2. Add icon for SisSceneCell which is "FILM" from Font Awesome. It is
needed for the style integrity: any other CuteElements have their
icons. SisScene's icon will indicate if the SisScene is primary or not
by the color of it's shadow.
3. Add PossibleMoves enumeration to SisSceneContextMenuBuilder to use
instead of integer values. This increases code readability.