Generate CuteNode names for hardcoded Projects
A name is a must have for CuteNode. When opening editor panel for
CuteNode from one of the hardcoded projects the name field does not
pass the validation, so "OK" and "Perform Test" buttons are not
enabled. And because of that, it's hard to test newly written
CuteNode-specific editor panels - the name should be fixed each time.

This patch makes name autogeneration for all hardcoded CuteProject's
CuteNodes.