Do not init() Element instantly on user input
CuteElement initialization is a resource hungry operation. Let's not
initialize instantly when GUI demands it, because the user might not
finished typing into fields. Instead let's play a text animation and
when the text animation will finish playing itself to the end, let's
finally do initialization. The animation can be interrupted by new user
input and started over shifting the moment of initialization to the
future.

This patch makes it so.