Fix memory leak with ValidationSupport
Before this patch ValidationSupport was accumulating registered
controls which prevented unneeded controls and views to be garbage
collected.

This patch brutally solves this problem by creating new
ValidationSupport object on each bind to new CuteElement. So old
controls, old views and old ValidationSupport objects can go straight
to hell as they are not linked anymore to the running code.