Reinitialize CuteProject on hitting Apply button
When the user has finished editing CuteNode (CuteElement) to fix it
from broken state and clicks "Apply" button, the parent nodes of
CuteNode might remain broken, because they still think their child is
still broken.

This patch fixes this problem by reinitializing the whole CuteProject.

But there's a drawback.
It is CPU intense operation and in the future it is better to:
- Run nice animation while CuteProject is reinitializing in
  non-GUI thread.
or
- Find a way to reinitialize only parents.