Fix CuteNodeContainer and bindings in GUI
CuteNodeContainer looks ugly and behaves ugly in GUI.
It doesn't have it's own icon in CuteTreeCell and also leads to crash.

This patch fixes it.
Also this patch fixes another bug that became visible after:
not proper unbinding from CuteElements in GUI. I was even forced to
read some documentation, it was that bad. So with this patch view
properties unbind properly from CuteElements.

Peace.