Fix IntegerTextField and DoubleTextField
When the numeric value of CuteElement's property equals "0" (zero),
and become bound with IntegerTextField's numberProperty, the text in
IntegerTextField stay's as empty. It's because ChangeListener on
IntegerTextField's numberProperty is not triggered after binding
(numberProperty and CuteElement's property have same value - "0").
And because it's not triggered, text in GUI's field is still empty.

This patch solves this problem not only in IntegerTextField, but also
in DoubleTextField.