Reinit not only whole project, but CuteElement
Problem: after fixing CuteElement in Element Editor panel, the
CuteElement might stay looking broken in the Tree.

Explanation: we reinitialize whole project, because parents might be
broken as well and fixing child might fix them also. But sometimes
reinitialization of current CuteElement might not occur. For example,
if Actor is broken because it doesn't have a Checker, the Actor will
not check it's Actions on reinitialization - it already knows it's
broken.

This patch makes Element Editor panel not only reinitialize whole
project, but also reinitialize particular CuteElement.