Fix VariableSetterAction not working
As UserVars.get(key)) may return null String, it's bad to put it first
in "equals" condition.

This patch fixes the problem by inverting equals condition in
VariableSetterAction.

Also it ensures that the first argument to equals method is non-null
by doing check in init(). It makes it also in VariableChecker.