Make validation decoration appear with animation
This patch makes validation decoration animated. Particularly it makes
it not just instantly appear on screen, but grow in size from 0% to
100%. The animation is initiated on first validation fail and
subsequent, but only if validation message changes. This is made
because the user might fix the current error in the field, but get into
new error. So when animation is played, the user will know that he
needs to reread validation message.

But it would be even nice if validation decoration would disappear also
with animation. But it would be hard to implement it.