Reinit CuteElement on every valid user input
Before this patch there was no reinitialization of CuteElement when the
user input was the same as CuteElement's initial property value, but it
changed back from invalid to valid.

This was bad, because CuteElement can become broken during user input,
but when it's property value is changed back, it doesn't reinit to
become healthy again.

This patch makes CuteElement to init() on every valid user input, not
only when this input differs from initial CuteElement's property value.