Fix reinit of CuteElement when fields are invalid
When field validation fails, CuteElement's init() should not be
invoked. But sometimes when init() is scheduled after finish of
"buttonsBeforeInitDotsAnima" animation in ElementEditorController,
the user can quickly change field value to invalid value while
animation is playing. Once it finishes playing, the CuteElement
reinitializes.

This patch fixes it by... refactoring of CuteButtonsStatesManager and
introducing it's new boolean property -
needToCancelScheduledCuteElementReinit which is meant to have
invalidation listener that cancels the animation.