Add INVALIDINPUT status for CuteElement's health
Sometimes when the user starts editing healthy CuteElement and provides
wrong input to editing fields, CuteElement still stated as healthy in
the left subpanel of Element Editor, which is not good. Even though
"OK" and "Apply" button are disabled at this moment, it's better to
give to the user an additional hint why this is happening and what
should he do.

This patch adds special health status for CuteElement -
ElementHealth.INVALIDINPUT which should be set exclusively in GUI. Now
the user will see "Awaiting for valid input" message instead of
CuteElement's health in such situations.