Make flexible behavior for user input restricting
Before this patch if CuteElement is not editable, the CuteController's
view just become disabled. But CuteController for CuteNodeContainer
doesn't have fields for input in it's view. Instead it provides
valuable information to the user. It's sad that this information is
grayed out.

This patch makes it possible not to disable the view completely, but
allow CuteControllers to override this default behavior.
CuteNodeContainer's view is now not disabled.