Allow empty names in some CuteElements
VariableSwitchAction requires it's CuteNodeContainers be able to have
empty name, so VariableSwitchAction can check a variable for empty
value. But it's not possible now in the program as nameTextField
validation fails on attempt to input an empty name.

This patch adds new property to ElementInfo - emptyNameAllowed, so
the GUI can know when it's okay to turn off validation that checks if
name is empty.