Implement SuperSoundManager
This patch implements SuperSoundManager that allows to control volume
and playback of multiple sounds.
This patch also substitutes SoundClip field with MediaPlayer in
SoundActions for the flexibility.

Thanks to these changes the "Stop all sounds" button is finally
working.
Also Volume slider now immediately affects the volume of all sounds
that are playing instead of affecting only the sounds that haven't
started playing yet.