Add networking support for Streaming Programs
This fat patch adds support for networking in StreamSis.
With this patch StreamSis will be able to communicate with Streaming
Programs.

Particularly it adds support to communicate with OBS Studio program
with websocket plugin turned on. It can get the name of the current
scene and set the current scene by name in this program.

In the future the other Streaming Programs will be supported. The patch
is written in a way to make implementations of their protocols easy
thanks to few abstraction layers.