Improve getting sick behavior of some elements
CuteElement can get sick when some conditions do not allow it to
perform well or when they make it's last execution useless.

But CuteElement should also be able to restore healthy status. And it
should occur on the first successful iteration of execution after
failed one.

This patch modifies the behavior of some CuteElements to make them
follow the described above behavior.