Fix Editor buttons not working on slow machines
If create new CuteElement in Structure view and quickly click on it
while Element Editor panel is opened, validation will fail continuously
even after providing valid input in this panel.

In short: this is because of race condition and this patch fixes it by
simply adding "synchronized" modifier to "connectToCuteElement" method.