Add descriptions for all CuteElements
The user needs to know how each particular CuteElement work.

This patch adds descriptions for CuteElements.

Each description is now shown in tooltips.
For example, in the Element Editor panel near "Type" field there is a
question mark icon which has a tooltip with type description.
A tooltip with description also appears now on top of corresponding
MenuItems for adding new CuteElements when the user right clicks on the
Structure View.