Allow adding new Actor on right click on Actor
Now to add a new Actor the user needs to right click on free space on
the "Actor list". This can be problematic if there are already some
Actors and there's no free space in the list.

This patch adds ability for the user to add new actor by right clicking
on the existing Actor in list. Such option will be in context menu.

The same solution was implemented for SisScenes.

TODO: implement same solution for CuteElements in TreeViews, but it
will be a bit harder.