Actor feature: sleep duration after success check
This patch adds ability for Actor to stop checking for a specified
amount of time after the first successful result in a row taken from
Checker.

For example, this is useful when RegionChecker has easy to lose
target. When such target appears, RegionChecker finds, loses and finds
again target, which is leading to executing On Actions twice.