Fix 'invalid input' after race condition
In some cases MultiTargetRegionChecker's GUI controller can't pass
validation during editing.

It is because MultiTargetRegionChecker's controller in the end of one
of it's animations can report invalid input to the
CuteButtonsStatesManager instance. And before this animation ends, the
user can start editing another CuteElement. The new CuteElement's GUI
controller is bound to the same instance of CuteButtonsStatesManager.
Once the animation has ended playing, a report about invalid input is
sent to the instance of CuteButtonsStatesManager causing the Element
Editor to show "invalid input" error, which can't be fixed by providing
valid input.

This patch makes ElementEditorController recreate
CuteButtonsStatesManager instead of reusing it for editing different
CuteElements.