Fix CuteTreeCell's non-null result crash
Sometimes StreamSis crashes with:
"RuntimeException: TargetImageWithActions can only have null as result".
or
"RuntimeException: CuteElementContainer can only have null as result".

This is because of race condition that happens because
"PLatform.runlater()" is widely used in the CuteTreeCell.

Best thing to deal with this kind of problem is to remove the checks
for non-null result value in CuteTreeCell.

This patch does removes such checks.