Fix crash on applying changes to CuteElement
The previous patch has fixed a problem where exceptions were not thrown
outside a task. After the problem was fixed, exception like this became apparent:

"The number of initialized Elements(10) doesn't match with the number
of all Elements(9) in the Project."

This happens when applying changes to CuteElement by clicking on
"Apply" or "OK" button. Long story short: the copy of the CuteElement
for editing was also counted by Project as initialized CuteElement.

This patch fixes the problem with dirty workaround.